Move private methods to partial class
